Lazy indexers now accept multiple transforms
(schwardt, tmauch) Provide the ability to add extra transforms to a
LazyIndexer after instantiation. The indexer now contains a list of
LazyTransform objects representing a transform chain. The LazyTransform
wraps the transform itself, its final dtype and shape transform together
with a new name field for identification. The transform is explicitly
restricted in the changes it can make to the data shape, as these are
meant to be local self-contained transforms such as delay correction and
gain calibration (but not averaging). Provide a __str__ printout to
clarify the transform structure. Turn the shape and dtype into properties
as they may now change dynamically as the chain is edited.
Add transforms to ConcatenatedLazyIndexer which apply on top of the
underlying indexer transforms to keep the same interface. While H5DataV2
is largely unaffected, the H5DataV1 indexers are greatly simplified.

git-svn-id: https://katfs.kat.ac.za/svnDS/code/katfile/trunk@16596 b56c7a7d-ac47-0410-a905-b6f7de78896a