Fix exception due to unknown select() kwarg
If an unknown / invalid keyword argument is passed to select(), warn
the user instead of raising an exception, as these kwargs may originate
from places like scape and the pipeline where they are mixed with other
unrelated kwargs. The old behaviour currently causes scape to crash.
It is still useful to warn the new user unfamiliar with select().