Print warning if logging not configured
If no logging is configured by the time the library logs something,
the log message is silently lost via a NullHandler. This is clearly
not useful to users. On the other hand, a Python library is not
allowed to configure the logging system itself as this should only
happen once at the highest application level and under user control.

In order to assist users who are not used to configuring logging
themselves but also want to know about log messages they might be
missing, the library now prints a warning message if no logging
has been configured and suggests a simple logging solution.