Revert to old DataSet.select() behaviour
This again raises an exception instead of logging a warning when
DataSet.select() encounters an unknown keyword argument, in effect
undoing the changes of PR #2. However, an added *strict* keyword
can be used to disable the exception. By default the exception is
enabled for the benefit of command-line users. Also improve
docstrings.