Fixed a problem with strict selection
The 'strict' keyword argument that is passed to :meth:select was not being
saved when alternative selections were being made inside a :class:DataSet object,
for example when calling the scans() method. This would cause a 'strict=False'
selection to be reset to True (the default) and raise errors when non-standard kwargs
are passed to :meth:select. Adding the 'strict' keyword to the list of
kwargs passed to 'select' and using kwargs.get(..) to get its value with a sensible
default solves the problem.