Add support for observation labels and parameters
The observation parameter dict is constructed by replaying the
obs_params sensor events while the observation label now has a
dedicated sensor on the Observation component instead of lurking
in the Markup group.

In addition, group the sensors in the cache according to the class
of the component where they came from. This makes the sensor
groups less arbitrary than before while keeping the same names.
('Antennas' still has to become 'AntennaPositioner' though.)