Fix loading of attribute-like sensors
Both obs_params and dbe_mode are attribute-like sensors that take
on values that apply to the entire observation and their actual
timestamps do not matter. In fact, several events may occur within
a single dump period and these values should be preserved and not
culled. The solution is to not extract the sensor to a CategoricalData
object but rather return a list of all its values for further use.