Remove quicklook option and deferred timestamps
Since the timestamp HDF5 dataset is now written as a contiguous block
at the end of the HDF5 file it is quick to load and therefore no longer necessary to defer the loading of timestamps. The quicklook mechanism
and lazy indexing of timestamps are now obsolete and we revert back to
loading the timestamps once upon opening the file.