Use eval to unpack obs_params values
The obs_params values were originally repr'ed to strings in the
CaptureSession before getting marshalled into a KATCP-like spead stream
by cam2spead. We therefore need eval() to get them back to their original
form. It's not as bad as it sounds, as we suppress the context (and
eval != exec).