Extend flux model frequency range to fit data
Some of the targets in the default source list (e.g. Hydra A) have flux models
that don't extend over the full KAT-7 frequency range. The valid frequency
range on these continuum models seem quite arbitrary and should ideally be
extended to the full instrument range.

After comparison with models derived from the Kuehr 1 Jy catalogue, it seems
reasonably safe to extend the flux model frequency ranges by a factor of less
than two (at least for Hyd A!). Warn the user as this might not be valid in
all circumstances. The implicit assumption is that the source is continuum
with a smooth spectrum. The real fix is to change the source list.