Rotate baseline labels and force time_scale
The RTS correlator has a bug (Nov 2014) whereby the bls_ordering labels
do not match up with the data in xeng_raw, but is rotated by one position.
Correct for this by default and also do the timestamp fixes by default.
Once the correlator is complying, the defaults can be changed back.