Base dummy sensor data on initial values instead
When a sensor has no values in the file, it is replaced by a dummy data set.
This dummy set now uses the initial_value property of the sensor as a filler
value and also sets its dtype to that of the initial value, if available.
Previously it only determined its value from the sensor dtype. In order to
make this happen, parse the sensor properties first before creating the
dummy data set.

This circumvents the problem of truly empty sensor datasets in the file
(not just empty after clean-up) which have the frequently incorrect default
dtype of np.float64. The main culprit is Observation/label, which should be
a string. The initial_value for this sensor is sensible, though.