Improve RTS noise diode sensor
Invert the selection process to turn raw sensor into boolean value, as
the RTS sensor actually contains the duty cycle as a float. Having a
duty of anything above 0.0 will now register as noise diode 'on'.

Use the escaped version of the KATCP sensor name, as this seems to be
the rule for v3 files.