Remove initial blank target
In RTS, the antenna sometimes remains in stop mode when capture starts,
which implies having no target (aka 'Nothing, special'). This typically
messes up the first compound scan. As a workaround, remove this spurious
blank target and replace it with the next (hopefully meaningful) target.
The 'stop' scan thereby gets grouped with the next proper set of scans.