Fix concatenation with obs_params
Some observation parameter values are unhashable. For example, nd_params has
a value like {'diode': 'coupler', 'off': 0, 'on': 0, 'period': -1}, and a dict
is unhashable. The standard way to find the unique values in a sequence is to
put them into a set but that only works if they are hashable. Sorting the sequence followed by applying the groupby function in itertools provides a way around this (provided the values are at least orderable / comparable).

See http://stackoverflow.com/a/2527608/942503 and the discussion around it.