Go back to itertools.groupby
The unique_in_order method cannot handle unorderable objects like None.
We actually don't need a list of unique values. We just want to know if all
the values are the same or not (in other words, whether they are in a single
group according to groupby). This does not require sorting.