Handle arrays when concatenating indexers
Up to now the timestamps, vis, weights and flags in each DataSet return
a LazyIndexer, and ConcatenatedLazyIndexer assumed that it will get a
sequence of such indexers. However, the H5DataV3 timestamps are returned in
a NumPy array, and more DataSets will potentially go this route in future.

Instead of forcing all DataSets to use LazyIndexers with their overhead,
fix ConcatenatedLazyIndexer to accept a sequence of indexers and/or arrays.
Any raw arrays will be wrapped in LazyIndexers to keep things simple.