Obtain centre frequency from obs_params
Contrary to popular belief, the RTS system supports different centre
frequencies, thanks to its inclusion of a Ku-band receiver. The CBF actually
does not know the sky centre frequency anymore (and doesn't need to).
The data proxy needs it for delay correction and therefore knows it, so one
option is to include its delay-center-frequency sensor. On the other hand, the
soon-to-be-updated RTS CaptureSession sets and gets the frequency from this
sensor already and then passes it to obs_params. Reading it from obs_params
is therefore the easiest solution.

The only restriction (and a justifiable one) is that there can only be a single
centre frequency per file. Since changing the centre frequency amounts to a new
receiver and/or a new correlator product, this is fair.