Identify the active receiver on each antenna
This adds a receivers mapping to the basic DataSet that maps antenna names
to the corresponding active receiver on each antenna. The receiver is
identified by its band indicator and its serial number (e.g. 'l.4').
If this information is missing for a specific antenna, that antenna is left
out of the dict. This is useful for identifying the corresponding noise diode
model.

The ConcatenatedDataSet follows the same approach as for obs_params: expand
the receiver identifier to a full list of identifiers if they differ between
data sets to be concatenated.

This addresses JIRA ticket MRTS-178.