Revert centre freq and allow override
It is risky to depend on a script option / obs_param to determine the
centre frequency. Currently the centre frequency is not set in RTS
observation scripts and it then falls back to the old KAT-7 default
of 1822 MHz, causing havoc.

Rather go back to the way it was, and allow centre_freq to be overridden
during analysis to deal with the funny receivers like Ku and hacked UHF.
The centre_freq parameter has units of Hz to fit in with the rest of katdal
but this might confuse script users who are probably used to the default
unit of MHz for scape and CaptureSession.

Once we have more bands and also a working narrowband product, the centre
freq calculation needs to be revisited and extended.