Add default values for indexer + rx serial numbers
This ensures that katdal does not crash when one of these sensors are in
the HDF5 file but empty (like in 1437386854.h5). Since the default h5 dataset
type is float, the serial numbers became NaN in the absence of any data.
The assignment of initial sensor values of the appropriate type fixes this.
The default band is therefore 'no band' while the default receiver serial
number indicates 'not installed', both safe choices.