Fix duplicate final dump indexing for vis et al
The H5v3 file format has timestamps as an actual array instead of a lazy
indexer. If the file contains a duplicate final dump, this array will be
one shorter than the corresponding dimension on the visibility dataset in the
file. This causes problems when the time_keep boolean flags are used to index
vis (and weights and flags).

After refactoring, the solution is to handle this mismatch in the common
vis-like indexer, where an extra False is appended to the time_keep flags so
that the duplicate final dump is always ignored. This seems better than the
alternative of forcing time_keep to track the underlying data shape and then
having to deal with the duplicate dump at the DataSet level.