Make 3-D vis / weights / flags optional
If squeeze=True is added when opening an H5v3 file, the vis / weights / flags
indexers will not force the output arrays to be three-dimensional. This is
more natural and aligned with indexing in NumPy. The 3-D restriction was aimed
at e.g. signal displays that need to track the dimensions of generic vis arrays
and unnecessarily complicates the standard use case.

The default value for squeeze is False for now to preserve backwards
compatibility, but this might change in future.