Improve reference antenna selection
In some instances the antennas in the CAM metadata stream and those involved
in the CBF stream differ (this is a bug). The default refant used to be the
first CAM antenna, which might not be in use in the current experiment. The
default refant is now the first of the antennas common to both CAM and CBF.