Ignore empty strings in obs_params sensor
The kattelmod simulator currently insists on initialising each sensor on
startup, even if no data is available. Since the obs_params sensor is of type string it is initialised with an empty string, which breaks the current loader.
The solution is to ignore empty strings when assembling the obs_params dict.