Override receiver band in H5DataV3
This adds the 'band' keyword argument which allows a manual override of
the receiver band on all dishes. This is useful if the finicky
ap_indexer_position sensor is acting up on some dishes by turning up
as 'undefined'. Its only effect is to correct the contents of the receivers
attribute, which is in turn used in scape to load the appropriate noise
diode models.

The 'band' is a proxy for the indexer position and not the actual receiver
and this therefore does not cater for receivers being mounted out of place
on the indexer.