Make observation script log a top-level attribute
The obs script_log is not a standard sensor in that you do not want to
interpolate its values onto the data timestamps. Instead, extract its
values verbatim, turn them into a list of strings (one per log line) and
make it a top-level attribute in the DataSet. This also provides a unified
way to access the log between KAT-7 and MeerKAT.