New TelstateSensorData class
This is a feeble attempt to expose telstate sensors within an HDF5 file in
the same manner as the usual sensors under TelescopeModel. It is feeble
because we should really be wrapping TelescopeState itself and let the
HDF5 telstate store be an incarnation of that. The two main things to take
care of is faking a 'status' column and unpickling the data upon request
(also, getting the dtype right).