Extend support to older versions of TelescopeState
Before 2016-05-09 the TelescopeState data in the HDF5 file contained the
str() representation of the sensor value as opposed to a pickled string or
even the repr(). It is therefore not generally decodable but a safe_eval
can handle the common cases of numbers, bools and strings. Even dicts seem
to work... If unsure what to do, return the string itself and hope for the
best.

The TypeError check is removed for now to see where it manifests itself
on archived data.