Provide safe version of linear interpolation
Add two safety measures in the absence of scikits.fitting:

  - If only a single data point is provided, do zeroth-order interpolation
    instead
  - Don't extrapolate linearly past the provided range of x values; instead,
    use the closest y value (another zeroth-order interpolation)

The first prevents NaNs and the second prevents crazy sensor values as
far as possible (i.e. if it is wrong, let it at least be feasible).

This does not affect the scikits.fitting routines, which already does the
first measure but not the second (and takes precedence if installed).