Make weights a property like vis
Rather use the usual select() to pick weights and make weights
compatible with the behaviour of vis. The base class simply assigns
the selected weights (as names) to the _weights_keep attribute. It is
then up to the specific format class to do something with the selection.

In the case of H5DataV3, look up the names in the table of known weights
(which currently and probably forever only has a single entry, namely
'precision'). As a shortcut, since there is only one weight type, either
return the weights or an equivalent number of ones if nothing is
selected.

The shape of the Data/weights dataset in the HDF5 file is now the same
as that of Data/correlator_data because there is only one weight type.
Since v3 files never had weights until this week, this is a safe
redesign. The weights_channel dataset still has to be incorporated in
the weights calculation.