Make flags a property like vis
Rather use the usual select() to pick flag types and make flags
compatible with the behaviour of vis. The base class simply assigns
the selected flag types (as names) to the _flags_keep attribute. It is
then up to the specific format class to do something with the selection.

In the case of H5DataV3, look up the names in the table of known flags
and turn them into a bit mask to be used in the flags property.
Streamline the reversal of the mask indices due to the np.packbits
MSB-first policy, so that known_flags stay in the natural order.
Also assert that there are 8 flag types, just in case. Anything else
would complicate matters in the bit allocation.