Give H5DataV2 the v3 vis/weights/flags interface
Since v2 and v3 have essentially the same vis + weights + flags
interface, copy the v3 interface across to v2. Also add the squeeze
parameter. Eventually this duplicate code will be spun out to a
common base class but that needs more thinking and we are impatient.
(The tricky bit is getting selections to pass into this base class
and distinguishing between formats and data containers).