Rename squeeze to keepdims and change default
Change the name of this parameter to match the corresponding NumPy
concept (see e.g. np.sum or np.mean). It has the opposite sense of
squeeze. In the process, change the default setting to not keep dims
but squeeze instead, as this is typically what the user intends
(while keepdims should be a special case).