Fix weight shape issues
The squeezing of the weight dimensions exposed some errors in the shape
handling code. Now always disable keepdims in the weights_channel
indexer, as the keepdims transform should only happen right at the end.

This way the high-res weights shape tuple will always be a prefix of
the lo-res weights shape tuple inside the extractor transform, and at
most needs an extra dimension at the end. This ensures proper NumPy
broadcasting when the two weights data sets are multiplied together.