Derive band from subarray instead of AP indexer
The band selected for the subarray (reflected in the band sensor
on the subarray controller) now determines the receiver / band name.
The AP indexer position is no good for this as receivers get moved
out of place and that sensor is also notoriously fickle, easily
becoming 'undefined'.

In addition, the selected band also determines the centre frequency
and sideband orientation for the spectral window (at least for
wideband). This also caters for custom receivers like the current 'x'
(actually Ku) and 'u' (actually a custom / hacked UHF). Eventually
these receivers will either identify themselves or be distinguishable
from the real McCoys.

This also allows a more accurate determination of the receiver serial
number for the purpose of obtaining e.g. noise diode models.