Soften stance on unknown centre freq
Rather set the centre frequency to 0 Hz if unknown. Then the file
can still be opened to inspect other things, like the vis data and
target info. The archiver will also very much prefer this...

This way, the user is warned and the frequency is obviously wrong,
which will hopefully avoid confusion in this case.