Add an output Measurement set option
If supplied, h5toms will use the filename schema
specified in the option to construct the output filename.
Otherwise, as previously, it will use the
input HDF5 filename to construct the output filename.

This commit provides users with the option to dump
the output Measurement Set in a different directory
to the input HDF5 files.