Try to autodetect band on old RTS too
RTS did not have the subarray band attribute / sensor before 2016-07-21.
It only had the AP indexer position per antenna. In this case guess the
receiver band as the most common valid indexer position amongst all
antennas in the subarray.

This is not very reliable as the indexer sensors are frequently invalid
and receivers may also be installed in the "wrong" indexer position.
Nevertheless try one's best and warn the user of the dangers, thereby
avoiding the user having to override the band for all old RTS files.