Invalidate centre freq on channel mismatch
Keep the original channel width though. This assumes that the data
channels are a subset of the original CBF channels (as they should
be in this use case). We don't know which subset though, so invalidate
the centre freq for now. The order of these spw_params settings are
quite significant by now.

In future, we *could* even guess the centre frequency from the
config['sdp']['ingest']['output_channels'] key in telstate...