Fix upside-down MeerKAT images
There are two major changes:

1. When the spectral window uses lower sideband, conjugate the
visibilities on load, because the correlator/ingest do not correct for
it. This brings all data to the MeerKAT sign convention for
visibility phase. For v1 and v2 it's hardcoded (since they hardcode for
lower sideband), while for v3 it is dependent on the band.

2. Change the sign of UVW coordinates to match the CASA convention.

For MeerKAT the second change will make images made via h5toms come out
the right way up. For KAT-7 the two changes will cancel out when
imaging, although the sign conventions will now match MeerKAT.

Fixes SR-811.