Make unique_values a list and wrap where needed
Change the unique_values attribute of CategoricalData from an ndarray
to a list, as this is the only way to preserve sensor values that are
arbitrary objects (like lists and ndarrays). This has several
repercussions throughout the code, like losing the ability to do
advanced indexing and math ufuncs and not having a built-in dtype.

In addition, ensure that the sensor values are hashable by wrapping
object values in a HashableValueWrapper on demand via the
_hashable_values property. Object values are assumed to be wrapped on
construction of CategoricalData and then unwrapped in unique_values.
Whenever comparisons are needed (e.g. in rich comparison operators,
when adding or removing data points and during construction from raw
sensor data), access _hashable_values instead and wrap ancillary data.
This probably has some performance hit for these operations but spares
the user from having to think about unwrapping objects.

Rewrite the unique_in_order function to use OrderedDict, which is about
6 times faster than np.unique without requiring data to be sorted, and
also unwrap objects when done. Simplify the rich comparison operators
of CategoricalData which now also work for wrapped objects (equality,
not ordering). The sensor dtype is now calculated on the fly. The
sensor_to_categorical function wraps initial values, greedy values and
transforms where needed.

Rewrap all objects unpacked from the HDF5 TelescopeState cache, as
the contents of this cache has changed over the years and cannot be
assumed to contain pickles. There is no more need for np.array_equal.