Introduce TelstateSensorData class
This class wraps raw (uninterpolated) sensor data stored in a
TelescopeState object. The data is only read out on item access and
then cached on the object. Any objects are wrapped to be hashable.

The caching should be fine as a SensorData object is typically replaced
by either a CategoricalData object or a NumPy array as part of sensor
extraction, right after the caching occurs.