HashableValueWrapper -> ComparableArrayWrapper
Abandon the use of pickles for hashing and get rid of the whole
edifice. Replace this with a lightweight wrapper that only improves
equality comparisons for ndarrays. It is now fast and convenient to
wrap all objects when doing CategoricalData add/remove/compare.
The _comparable_values property is therefore faster than the
_hashable_values it replaces and does not depend on dtype anymore.

The unique_in_order function falls back to list-based lookup instead
of dict-based lookup (the whole reason for hashability in the first
place), which is slower but safer. Since there are relatively few
object-valued sensors with a small number of events each, this should
not matter.