Major rework of sensor dtype machinery
Sensors stored in telstate have an issue: they are pickled in raw form.
Their data types are therefore obscured until the first value is
extracted and unpickled. In fact, each event of the sensor could even
have a completely different dtype, but let's not go there... We prefer
to preserve the illusion that each sensor has a unique dtype.

The dtype is needed for two things: filling in missing data when
concatenating sensor caches from two separate datasets, and deciding
how to interpolate the sensor data onto the data timestamps. The latter
happens after extraction but the former is needed during initialisation.
The real problem is that dtype inference is potentially slow as it
requires at least a mini-extraction by itself.

The solution is to defer specifying the dtype until after extraction.
Set the dtype to None if the type is still unknown (coincidentally also
a valid dtype implying 'figure it out from the data'). The main work
is to deal with None dtypes in the concatenated case. Here we use the
SensorData base class as a marker that sensor data is both missing and
with unknown type in a sensor cache.

As suggested by Bruce, centralise the dtype inference in a single
function to clean things up. This also fixes a bug in replacing
missing sensors with dummy data in ConcatenatedSensorCache, which only
ever fixed the last cache as opposed to all of them.