Change infer_dtypes to common_dtype
After discovering np.result_type, rewrite common_dtype to use this.
This now infers the common dtype that results from np.concatenate,
which clarifies the function's purpose. Also clarify the description
of infer_dtype.

This also stops pretending that a unique dtype is being enforced in
ConcatenatedSensorData. The truth is that the latest telstate-based
sensors start off with a dtype of None (i.e. unknown dtype), which
makes every sensor compatible. Then, when the dtypes are discovered
in SensorData.get() they are only potentially validated in the case
of missing sensor data, since we are avoiding dtype inference as far
as possible. Rather embrace the fact that we are very type-loose due
to our use of pickles and let the user bear the consequences...