Improve dump -> event parsing
Rework the sensor_to_categorical function in response to the popularity
of longer dump periods (e.g. 8 seconds) on MeerKAT. The existing
function works OK if the dumps are shorter than the typical dish
activities like tracks and slews. With 8-second dumps there is the
problem that short slews (as found in e.g. interferometric pointing)
may start and end within a single dump. This mini-slew is then filtered
out by the dump -> event parser, resulting in a single track containing
multiple pointings.

The new version of the function associates an entire dump with a greedy
event, even if there is only a whiff of that event within the dump.
Any non-greedy events at the end of such a dump is carried over to
the next dump in this scenario. This logic is split out to its own
function (_single_event_per_dump).

Other minor enhancements include the culling of sensor events outside
the dump timestamp range (useful for big multi-observation telstates)
and a reordering of steps so that sensor values are transformed first.