Revert selected categorical sensor to array form
As part of the recent upgrade to categorical sensors it was deemed
more efficient to return a list of sensor values when selection is
enabled as opposed to an array, as the former duplicates pointers
to values while the latter duplicates the values themselves if they
become amalgamated into a multi-dimensional array.

In the end it turns out that a few user scripts assumed the array
output form (by e.g. doing arithmetic with the return value). This
is especially true for categoricals like bool and int. The onerous
case of the multi-dimensional array-valued sensor is now reconsidered
to be the exception instead of the norm and we revert to the original
array form for all selected sensor outputs from the sensor cache.

It is now the user's problem if a multi-dimensional array-valued sensor
is selected without considering its memory requirements.