Get rid of NumPy FutureWarning
Since categorical sensor values are compared by value and not by
identity when doing a == b, it triggers a NumPy FutureWarning:

  numpy equal will not check object identity in the future.
  The comparison did not return the same result as suggested
  by the identity (`is`)) and will change.

This is because the current NumPy implementation (1.12.1 at least)
turns

  np.array([a], dtype=object) == np.array([b], dtype=object)

into

  np.array([a is b or a == b], dtype=bool)

while we are expecting something like the future NumPy behaviour of

  np.array([a == b], dtype=bool)

This commit does not change the result of the evaluation but removes
the warning, while also reducing the code clutter a bit.

See https://github.com/numpy/numpy/issues/4685 for more details.