Fix NumPy 1.13 error (boolean index mismatch)
Since NumPy 1.13 it is an error if an advanced index in the form of
a sequence of bools has a different length than the array it indexes.
In earlier versions this was a VisibleDeprecationWarning stating

  boolean index did not match indexed array along dimension X;
  dimension is M but corresponding boolean dimension is N [!= M]

This is brought on by the fact that the events sequence is always one
longer than the indices sequence in CategoricalData, storing the
total data length in the final events slot. The indexing was only meant
to happen on the first N-1 slots and did so implicitly until now.

This addresses JIRA ticket SR-877.