Fix categorical sensor parser
The `_single_event_per_dump` filter is supposed to mutate the events
array to simplify the process of pushing forward displaced events.
Unfortunately the code in `sensor_to_categorical` passed it a copy
of `events`, which defeats the whole purpose of the mutation.

Pass in the real `events`, slap on more warnings about the mutation
and even add a unit test to verify the finicky sensor->categorical
conversion.