Support standard UHF receivers
The standard UHF receivers have a digitiser sampling at 1088 MHz with
an unflipped bandwidth of 544 MHz in the second Nyquist zone
(544 - 1088 Mhz), resulting in a centre frequency of 816 MHz.

Delegate the old "fake" UHF receiver to the corner cases (only active
if the bandwidth is 856 MHz, i.e. it uses an L-band digitiser). Move
the calculation of bandwidth and channels earlier to facilitate this.