Prefer new-style noise diode sensors
The latest noise diode sensors have the band identifier as part of
their name. For example, the old L-band 'dig_noise_diode' is now
'dig_l_band_noise_diode'. This caters for the different origins of
the sensors as not all digitisers are served by the MeerKAT DMC.

Modify the main alias `nd_coupler` if a new-style noise diode sensor
was found (else remain with the default 'dig_noise_diode'). This should
have been a sensor alias, but that machinery can only change suffixes
and we need to move from TelescopeState to Antennas as well. Aliases
are also set up front and we need to add an alias a while later in the
initialiser.

Also ensure that all sensors ending in `noise_diode` are interpreted
correctly.

This addresses JIRA ticket MKAT-99.