Handle output L0 stream attributes
This is to handle the new attributes written by ingest
(ska-sa/katsdpingest#194) and filewriter (ska-sa/katsdpdata#71), which
allow for
- The file to contain either the continuum or spectral L0 stream
  (disambiguated by an HDF5 attribute on the /Data group);
- The file to contain only a subset of channels, and thus have separate
  n_chans, bandwidth, centre frequency to the CBF.

An area that will need some review is the precedence of the different
ways of determining centre frequency. Currently I've made it trust the
SDP-written value (e.g. sdp_l0_center_freq) over anything else, because
values determined from the band don't take account of channel subsetting
in ingest. However, that in turn means trusting the CAM centre
frequency, which is probably not correct for Ku band. I propose that the
right way to handle that is to fix CAM to be Ku band aware before trying
to update RTS with the latest ingest and filewriter.