Relax expected dump check
The SDP ingest emits a partial dump at the end of a capture which
consists of the leftover CBF mini-dumps. The last timestamp might
therefore not be exactly one dump period after the second-to-last one,
unnecessarily triggering this warning about excepted dumps being out.
The solution is to use the second-to-last timestamp in the check.

PS In the multi-ingest future the last dump will be aligned with the
rest again, so ignore this fix for v4.