Let DataSource aggregate metadata + data
This replaces multiple inheritance with aggregation. DataSource is
now the first stop for custom combinations of metadata and data.
Vis + flags + weights are always assumed to be there, whereas
previously some of them could be None. Now the entire data attribute
will be None if no data is available. This also simplifies the
VisDataV4 API.