Remove S3 access and secret keys
These are basically the username + password to the object store, so
don't require the user to provide it to the script and don't store it
in telstate. Also don't store the URL as that is an implementation
detail of the archive.

Instead, let the standard AWS config options provide the credentials
(e.g. ~/.aws/credentials or the relevant environment variables).
The boto3 / botocore library currently has no support for getting the
URL from config, otherwise I would have removed that too.