Store NumPy metadata in object
This is the next iteration of h5toobj. The shape, dtype and order
of each NumPy array chunk is stored as S3 object metadata or librados
xattrs, which allows the reading of the chunk to be standalone.
Use the existing NPY file header for this purpose. It's still better
than writing the array directly as an NPY file as that will require
a copy of the data. By wrapping the NumPy buffer in a file object,
this extra copy is now also avoided in the case of S3 (inspired by
cottoncandy).