Improved NumPy metadata representation
Instead of the messy base64-encoded binary string to be read by the
NPY header code to get the entire bit of metadata, store the dtype and
shape as separate keys in the object metadata (which is already a
dict by itself). The strings are handled with repr + np.safe_eval,
just like the NPY format code does. This bypasses the need for base64
and json, both with its own issues. The fortran_order flag is dropped
since the ordering is highly unlikely to be 'F'.