The ChunkStore in various incarnations
A *chunk* is a simple (i.e. unit-stride) slice of an N-dimensional array
known as its *parent array*. The array is identified by a string name,
while the chunk within the array is identified by a sequence of slice
objects which may be used to extract the chunk from the array. The array
is a :class:`numpy.ndarray` object with an associated *dtype*.

The basic idea is that the chunk store contains multiple arrays
addressed by name. The list of available arrays and all array metadata
(shape, chunks and dtype) are stored elsewhere. The metadata is used to
identify chunks, while the chunk store takes care of storing and
retrieving bytestrings of actual chunk data. These are packaged back
into NumPy arrays for the user.

There are four incarnations:

  - DictOfArraysChunkStore: works on dicts of NumPy arrays or HDF5 files
  - S3ChunkStore: works via the S3 object storage API (chunk == object)
  - RadosChunkStore: works via the Ceph RADOS API (chunk == object)
  - NpyFileChunkStore: works on a directory of NPY files (chunk == file)

Still missing:

  - An appropriate exception if a chunk is not found