Rework Rados store to use keyring and test it
The call to rados.Rados is updated to reflect reality. The chunk store
now accepts Ceph configuration either via config file or via dict, and
the keyring can be specified independently (a good idea).

Add the first unit tests for RadosChunkStore. This assumes that we
can connect to the seekat cluster, which limits the usefulness of
this test but is a practical choice since there is no "fake rados".

Also cater for S3 unit tests that may fail if botocore is not installed.