Make dtype non-optional
This normalises the interface of the various chunk stores. If the
requested dtype differs from the underlying dtype (when available),
raise a ValueError. In the process, move the float32 -> complex64
conversion needed for HDF5 files to h5toobj.py where it belongs.

Remove unneeded dask graph node (the keys of each chunk are really
arbitrary and do not need to refer back to the underlying dataset).