Improvements to low-level interfaces
For NpyFileChunkStore, disallow pickles / object arrays to improve
security. Also improve the creation of intermediate subdirectories.

For S3ChunkStore, stop pretending that we are avoiding a copy of the
array on the way to and from the wire. Always convert to bytes string
first. Also make stream reads more robust via closing context manager.