Fix opening of older v3 files
On pre-telstate v3 files, the _get_telstate_attr() helper function
crashes when it looks for improved versions of the dump period
attribute in the sensor cache. The problem is that the sensor cache
does not exist this early on in the metadata parsing... Fall back to
the default obtained from the TelescopeModel on this exception.