Use custom exceptions
Instead of the blanket `OSError`, split the exceptions into
`StoreUnavailable` and `ChunkNotFound`, which are logically different
problems. Centralise the various `_convert_xxx_errors` functions in the
base class and specify the conversion via an error map provided at
store initialisation. Since we are now responding to specific errors
as opposed to their base classes as well, the S3 botocore errors had to
change to their dynamic counterparts. Introduce standardised errors in
DictChunkStore as well.